Fix browser unit tests (#925)
* fix(wip): browser tests


* fix: remove commented out checksum tests as they've been moved into a separate file


* fix: update github actions to run unit tests in browser instead of integration


* fix: re-add `test/.eslintrc.cjs` file and fix formatting


* fix: add `DelegateContractId` to exports
