+++ index.html
@@ -3,16 +3,65 @@
     <head>
         <meta charset="utf-8" />
 
-        <title>Hedera™ JavaScript SDK</title>
+        <title>Hedera™ JavaScript SDK > Tests</title>
 
         <meta name="viewport" content="width=device-width, initial-scale=1.0" />
 
         <meta
             http-equiv="Content-Security-Policy"
-            content="script-src 'unsafe-inline' localhost:9001 localhost:9002 localhost:9003 localhost:9004 localhost:9011 localhost:9012 localhost:9013 localhost:9014"
+            content="script-src 'self' 'unsafe-inline' localhost:9001 localhost:9002 localhost:9003 localhost:9004 localhost:9011 localhost:9012 localhost:9013 localhost:9014"
         />
+
+        <link rel="stylesheet" href="node_modules/mocha/mocha.css" />
+        <base href="/">
     </head>
     <body>
-        <script type="module" src="./src/browser.js"></script>
+        <div id="mocha"></div>
+
+        <script src="node_modules/chai/chai.js"></script>
+        <script src="node_modules/mocha/mocha.js"></script>
+        <script src="node_modules/mocha-webdriver-runner/dist/mocha-webdriver-client.js"></script>
+        <script src="node_modules/mocha-webdriver-runner/dist/mocha-webdriver-client.js"></script>
+
+        <script class="mocha-init">
+            mocha.setup("bdd");
+            mocha.checkLeaks();
+        </script>
+
+        <script class="chai-init">
+            window.expect = chai.expect;
+        </script>
+
+        <!-- NOTE: all unit tests need to be imported here -->
+        <script type="module" src="test/unit/AccountAllowanceAdjustTransaction.js"></script>
+        <script type="module" src="test/unit/AccountAllowanceApproveTransaction.js"></script>
+        <script type="module" src="test/unit/AccountId.js"></script>
+        <script type="module" src="test/unit/ChecksumValidation.js"></script>
+        <script type="module" src="test/unit/ContractFunctionParameters.js"></script>
+        <script type="module" src="test/unit/DelegateContractId.js"></script>
+        <script type="module" src="test/unit/EcdsaPrivateKey.js"></script>
+        <script type="module" src="test/unit/EntityIdHelper.js"></script>
+        <script type="module" src="test/unit/Executable.js"></script>
+        <script type="module" src="test/unit/Hbar.js"></script>
+        <script type="module" src="test/unit/keccak256.js"></script>
+        <script type="module" src="test/unit/LedgerId.js"></script>
+        <script type="module" src="test/unit/ManagedNodeAddress.js"></script>
+        <!-- <script type="module" src="test/unit/NodeClient.js"></script> -->
+        <script type="module" src="test/unit/PublicKey.js"></script>
+        <script type="module" src="test/unit/Serialize.js"></script>
+        <script type="module" src="test/unit/Timestamp.js"></script>
+        <script type="module" src="test/unit/TokenPauseTransaction.js"></script>
+        <script type="module" src="test/unit/TokenUnpauseTransaction.js"></script>
+        <script type="module" src="test/unit/TopicMessageSubmitTransaction.js"></script>
+        <script type="module" src="test/unit/Transaction.js"></script>
+        <script type="module" src="test/unit/TransactionId.js"></script>
+        <script type="module" src="test/unit/TransferTransaction.js"></script>
+        <script type="module" src="test/unit/util.js"></script>
+
+
+
+        <script type="module" class="mocha-run">
+            mocha.run();
+        </script>
     </body>
 </html>
