+++ package.json
@@ -69,9 +69,9 @@
         "test:browser:chrome": "run-s test:unit:browser:chrome test:integration:browser:chrome",
         "test:browser:firefox": "run-s test:unit:browser:firefox test:integration:browser:firefox",
         "test:unit:browser": "run-s test:unit:browser:*",
-        "test:unit:browser:chrome": "vite -c vite.config.cjs serve --port 9001 test/ & sleep 10; mocha-webdriver-runner --headless-chrome http://localhost:9001/unit.html",
-        "test:unit:browser:firefox": "vite -c vite.config.cjs serve --port 9002 test/ & sleep 10; mocha-webdriver-runner --headless-firefox http://localhost:9002/unit.html",
-        "test:unit:browser:safari": "vite -c vite.config.cjs serve --port 9003 test/ & sleep 10; mocha-webdriver-runner --safari http://localhost:9003/unit.html",
+        "test:unit:browser:chrome": "vite -c vite.config.cjs serve --port 9001 . & sleep 1; mocha-webdriver-runner --headless-chrome http://localhost:9001",
+        "test:unit:browser:firefox": "vite -c vite.config.cjs serve --port 9002 . & sleep 1; mocha-webdriver-runner --headless-firefox http://localhost:9002",
+        "test:unit:browser:safari": "vite -c vite.config.cjs serve --port 9003 . & sleep 1; mocha-webdriver-runner --safari http://localhost:9003",
         "test:integration:browser": "run-s test:integration:browser:*",
         "test:integration:browser:chrome": "vite -c vite.config.cjs serve --port 9011 test/ & sleep 10; mocha-webdriver-runner --headless-chrome http://localhost:9011/integration.html",
         "test:integration:browser:firefox": "vite -c vite.config.cjs serve --port 9012 test/ & sleep 10; mocha-webdriver-runner --headless-firefox http://localhost:9012/integration.html",
@@ -99,6 +99,7 @@
         "@types/chai": "^4.3.0",
         "@types/crypto-js": "^4.1.0",
         "@types/long": "^4.0.1",
+        "@types/mocha": "^9.1.0",
         "@types/node": "^17.0.13",
         "@types/utf8": "^3.0.1",
         "@typescript-eslint/eslint-plugin": "^5.10.1",
