+++ src/exports.js
@@ -39,6 +39,7 @@ export { default as CustomFee } from "./token/CustomFee.js";
 export { default as CustomFixedFee } from "./token/CustomFixedFee.js";
 export { default as CustomFractionalFee } from "./token/CustomFractionalFee.js";
 export { default as CustomRoyaltyFee } from "./token/CustomRoyaltyFee.js";
+export { default as DelegateContractId } from "./contract/DelegateContractId.js";
 export { default as ExchangeRate } from "./ExchangeRate.js";
 export { default as FileAppendTransaction } from "./file/FileAppendTransaction.js";
 export { default as FileContentsQuery } from "./file/FileContentsQuery.js";
@@ -73,7 +74,6 @@ export { default as SystemDeleteTransaction } from "./system/SystemDeleteTransac
 export { default as SystemUndeleteTransaction } from "./system/SystemUndeleteTransaction.js";
 export { default as Timestamp } from "./Timestamp.js";
 export { default as TokenAllowance } from "./account/TokenAllowance.js";
-export { default as TokenNftAllowance } from "./account/TokenNftAllowance.js";
 export { default as TokenAssociateTransaction } from "./token/TokenAssociateTransaction.js";
 export { default as TokenBurnTransaction } from "./token/TokenBurnTransaction.js";
 export { default as TokenCreateTransaction } from "./token/TokenCreateTransaction.js";
@@ -86,6 +86,7 @@ export { default as TokenId } from "./token/TokenId.js";
 export { default as TokenInfo } from "./token/TokenInfo.js";
 export { default as TokenInfoQuery } from "./token/TokenInfoQuery.js";
 export { default as TokenMintTransaction } from "./token/TokenMintTransaction.js";
+export { default as TokenNftAllowance } from "./account/TokenNftAllowance.js";
 export { default as TokenNftInfo } from "./token/TokenNftInfo.js";
 export { default as TokenNftInfoQuery } from "./token/TokenNftInfoQuery.js";
 export { default as TokenPauseTransaction } from "./token/TokenPauseTransaction.js";
