--- test/integration.html
@@ -1,45 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-    <head>
-        <meta charset="utf-8" />
-
-        <title>Hedera™ JavaScript SDK > Tests</title>
-
-        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-
-        <meta
-            http-equiv="Content-Security-Policy"
-            content="script-src 'unsafe-inline' localhost:9001 localhost:9002 localhost:9003 localhost:9004 localhost:9011 localhost:9012 localhost:9013 localhost:9014"
-        />
-
-        <link rel="stylesheet" href="node_modules/mocha/mocha.css" />
-    </head>
-    <body>
-        <div id="mocha"></div>
-
-        <script src="node_modules/chai/chai.js"></script>
-        <script src="node_modules/mocha/mocha.js"></script>
-        <script src="./node_modules/mocha-webdriver-runner/dist/mocha-webdriver-client.js"></script>
-        <script src="./node_modules/long/dist/long.js"></script>
-
-        <script class="mocha-init">
-            mocha.setup({ ui: "bdd", timeout: "3600s" });
-            mocha.checkLeaks();
-        </script>
-
-        <script class="chai-init">
-            window.expect = chai.expect;
-        </script>
-
-        <!-- NOTE: all unit tests need to be imported here -->
-        <script type="module" src="integration/AccountBalanceIntegrationTest.js"></script>
-        <script type="module" src="integration/ClientIntegrationTest.js"></script>
-        <script type="module" src="integration/TokenCreateIntegrationTest.js"></script>
-        <script type="module" src="integration/TransactionIntegrationTest.js"></script>
-        <script type="module" src="integration/TransactionResponseTest.js"></script>
-
-        <script type="module" class="mocha-run">
-            mocha.run();
-        </script>
-    </body>
-</html>
