+++ test/unit/AccountAllowanceAdjustTransaction.js
@@ -1,4 +1,3 @@
-import "mocha";
 import { expect } from "chai";
 
 import Long from "long";
