+++ test/unit/AccountId.js
@@ -1,7 +1,8 @@
-import AccountId from "../../src/account/AccountId.js";
+import { expect } from "chai";
+
+import { AccountId, PublicKey } from "../../src/exports.js";
+
 import BigNumber from "bignumber.js";
-import { Client } from "../integration/client/NodeIntegrationTestEnv.js";
-import PublicKey from "../../src/PublicKey.js";
 
 describe("AccountId", function () {
     it("constructors", function () {
@@ -66,42 +67,6 @@ describe("AccountId", function () {
         expect(accountId.num.toNumber()).to.eql(25050);
         expect(accountId.realm.toNumber()).to.eql(50);
         expect(accountId.shard.toNumber()).to.eql(10);
-    });
-
-    it("should parse mainnet ID with checksum {0.0.123-vfmkw}", function () {
-        const accountId = AccountId.fromString("0.0.123-vfmkw");
-
-        expect(accountId.num.toNumber()).to.eql(123);
-        expect(accountId.realm.toNumber()).to.eql(0);
-        expect(accountId.shard.toNumber()).to.eql(0);
-
-        expect(accountId.toStringWithChecksum(Client.forMainnet())).to.be.eql(
-            "0.0.123-vfmkw"
-        );
-    });
-
-    it("should parse testnet ID with checksum {0.0.123-rmkyk}", function () {
-        const accountId = AccountId.fromString("0.0.123-rmkyk");
-
-        expect(accountId.num.toNumber()).to.eql(123);
-        expect(accountId.realm.toNumber()).to.eql(0);
-        expect(accountId.shard.toNumber()).to.eql(0);
-
-        expect(accountId.toStringWithChecksum(Client.forTestnet())).to.be.eql(
-            "0.0.123-rmkyk"
-        );
-    });
-
-    it("should parse previewnet ID with checksum {0.0.123-ntjly}", function () {
-        const accountId = AccountId.fromString("0.0.123-ntjly");
-
-        expect(accountId.num.toNumber()).to.eql(123);
-        expect(accountId.realm.toNumber()).to.eql(0);
-        expect(accountId.shard.toNumber()).to.eql(0);
-
-        expect(
-            accountId.toStringWithChecksum(Client.forPreviewnet())
-        ).to.be.eql("0.0.123-ntjly");
     });
 
     it("should parse 0.0.0", function () {
