+++ test/unit/ContractId.js
@@ -1,6 +1,5 @@
-import "mocha";
-
 import { expect } from "chai";
+
 import { ContractId } from "../../src/exports.js";
 import * as hex from "../../src/encoding/hex.js";
 import Long from "long";
