+++ test/unit/EcdsaPrivateKey.js
@@ -1,4 +1,6 @@
-import PrivateKey from "../../src/PrivateKey.js";
+import { expect } from "chai";
+
+import { PrivateKey } from "../../src/exports.js";
 import * as hex from "../../src/encoding/hex.js";
 
 const RAW_KEY =
