+++ test/unit/TransferTransaction.js
@@ -1,14 +1,16 @@
 import { expect } from "chai";
 
-import TransferTransaction from "../../src/account/TransferTransaction.js";
-import HbarUnit from "../../src/HbarUnit.js";
-import Hbar from "../../src/Hbar.js";
-import TokenId from "../../src/token/TokenId.js";
-import AccountId from "../../src/account/AccountId.js";
-import Transaction from "../../src/transaction/Transaction.js";
-import TransactionId from "../../src/transaction/TransactionId.js";
-import Timestamp from "../../src/Timestamp.js";
-import NftId from "../../src/token/NftId.js";
+import {
+    TransferTransaction,
+    HbarUnit,
+    Hbar,
+    TokenId,
+    AccountId,
+    Transaction,
+    TransactionId,
+    Timestamp,
+    NftId,
+} from "../../src/exports.js";
 import Long from "long";
 
 describe("TransferTransaction", function () {
