+++ vite.config.cjs
@@ -27,15 +27,20 @@ module.exports = {
             // note that this is NOT needed when consuming this package as the browser field in package.json
             // will take care of this
             "../src/index.js": "../src/browser.js",
+            "../../../src/encoding/hex.js": "../../../src/encoding/hex.browser.js",
             "../../src/encoding/hex.js": "../../src/encoding/hex.browser.js",
             "../src/encoding/hex.js": "../src/encoding/hex.browser.js",
+            "src/encoding/hex.js": "src/encoding/hex.browser.js",
             "../encoding/hex.js": "../encoding/hex.browser.js",
+            "./encoding/hex.js": "./encoding/hex.browser.js",
             "../src/encoding/utf8.js": "../src/encoding/utf8.browser.js",
+            "../../src/encoding/utf8.js": "../../src/encoding/utf8.browser.js",
             "../encoding/utf8.js": "../encoding/utf8.browser.js",
             "../src/cryptography/sha384.js": "../src/cryptography/sha384.browser.js",
             "../cryptography/sha384.js": "../cryptography/sha384.browser.js",
             "./client/NodeIntegrationTestEnv.js": "./client/WebIntegrationTestEnv.js",
             "../integration/client/NodeIntegrationTestEnv.js": "../integration/client/WebIntegrationTestEnv.js",
+            "../../src/client/NodeClient.js": "../../src/client/WebClient.js",
         },
     },
 };
